+++ src/main/java/com/example/domain/model/gate/Transition.java
@@ -1,5 +1,8 @@
-package com.example.domain.type.state;
+package com.example.domain.model.gate;
 
+/**
+ * 遷移前、遷移後の状態とイベントの組み合わせを扱う
+ */
 public class Transition {
     State from;
     State to;
