+++ src/main/java/com/example/domain/model/gate/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * ゲートの開閉状態のモデル
+ */
+package com.example.domain.model.gate;
