--- src/main/java/com/example/domain/type/state/package-info.java
@@ -1,4 +1,0 @@
-/**
- * 状態を扱う値オブジェクト
- */
-package com.example.domain.type.state;
